+++ src/Services/DynamoDb.php
@@ -1,13 +1,13 @@
 <?php
+
 namespace DreamFactory\Core\Aws\Services;
 
 use Aws\DynamoDb\DynamoDbClient;
 use DreamFactory\Core\Aws\Database\Schema\DynamoDbSchema;
 use DreamFactory\Core\Aws\Resources\DynamoDbTable;
-use DreamFactory\Core\Exceptions\InternalServerErrorException;
 use DreamFactory\Core\Database\Resources\DbSchemaResource;
 use DreamFactory\Core\Database\Services\BaseDbService;
-use DreamFactory\Core\Utility\Session;
+use DreamFactory\Core\Exceptions\InternalServerErrorException;
 
 /**
  * DynamoDb
@@ -29,7 +29,7 @@ class DynamoDb extends BaseDbService
             'class_name' => DbSchemaResource::class,
             'label'      => 'Schema',
         ],
-        DynamoDbTable::RESOURCE_NAME  => [
+        DynamoDbTable::RESOURCE_NAME    => [
             'name'       => DynamoDbTable::RESOURCE_NAME,
             'class_name' => DynamoDbTable::class,
             'label'      => 'Table',
@@ -54,21 +54,21 @@ class DynamoDb extends BaseDbService
 
         // statically assign our supported version
         $this->config['version'] = '2012-08-10';
-        if (isset($this->config['key']))
-        {
+        if (isset($this->config['key'])) {
             $this->config['credentials']['key'] = $this->config['key'];
         }
-        if (isset($this->config['secret']))
-        {
+        if (isset($this->config['secret'])) {
             $this->config['credentials']['secret'] = $this->config['secret'];
         }
 
         // set up a default table schema
-        $parameters = (array)array_get($this->config, 'parameters');
-        Session::replaceLookups($parameters);
+//        $parameters = (array)array_get($this->config, 'parameters');
 //        if (null !== ($table = array_get($parameters, 'default_create_table'))) {
 //            $this->defaultCreateTable = $table;
 //        }
+
+        $this->setConfigBasedCachePrefix(array_get($this->config, 'credentials.key') .
+            array_get($this->config, 'region') . ':');
     }
 
     protected function initializeConnection()
@@ -77,8 +77,6 @@ class DynamoDb extends BaseDbService
             $this->dbConn = new DynamoDbClient($this->config);
             /** @noinspection PhpParamsInspection */
             $this->schema = new DynamoDbSchema($this->dbConn);
-            $this->schema->setCache($this);
-            $this->schema->setExtraStore($this);
         } catch (\Exception $ex) {
             throw new InternalServerErrorException("AWS DynamoDb Service Exception:\n{$ex->getMessage()}",
                 $ex->getCode());